fix(security): disable `url` param feature by default
The `url` query parameter support fetch a specific `yaml` file
from remote by just appending `url=${path to yaml}` in the address bar

This feature might cause "spec faking" attack and we'd like to disable it
by default. A plugin is added in this commit to remove the `url` param
before swagger detect and fetch it.

You can use following url to test this behavior:
http://localhost:8003/default/documentation/petstore?url=http://localhost:8003/default/evil.yaml
It should render the correct spec instead of throwing an error message

Ref:
https://github.com/swagger-api/swagger-ui/security/advisories/GHSA-qrmm-w75w-3wpx

This fix FTI-5349